+++ src/javascript/ZeroClipboard/event.js
@@ -18,6 +18,12 @@ ZeroClipboard.Client.prototype.receiveEvent = function (eventName, args) {
   // special behavior for certain events
   switch (eventName) {
   case 'load':
+    // If the flash version is less than 10, throw event.
+    if (args && parseFloat(args.flashVersion.replace(",", ".").replace(/[^0-9\.]/gi, '')) < 10) {
+      this.receiveEvent("onWrongFlash", { flashVersion: args.flashVersion });
+      return;
+    }
+
     this.htmlBridge.setAttribute("data-clipboard-ready", true);
     break;