+++ test/client.js
@@ -4,6 +4,11 @@ require("./env")
 
 exports.client = {
 
+  tearDown: function (callback) {
+    navigator.mimeTypes["application/x-shockwave-flash"] = undefined;
+    callback();
+  },
+
   "Client without selector doesn't have element": function (test) {
 
     var zeroClipboard = require("../ZeroClipboard"),
@@ -26,6 +31,31 @@ exports.client = {
     test.ok(clip.element);
     test.ok(clip.htmlBridge);
     test.ok(clip.handlers);
+
+    test.done();
+  },
+
+  "Detecting no flash": function (test) {
+
+    var zeroClipboard = require("../ZeroClipboard"),
+        clip = new zeroClipboard.Client();
+
+    // Test that we don't have flash
+    test.equal(zeroClipboard.detectFlashSupport(), false);
+
+    test.done();
+  },
+
+  "Detecting has flash mimetype": function (test) {
+
+    var zeroClipboard = require("../ZeroClipboard"),
+        clip = new zeroClipboard.Client();
+
+    // We're faking it here.
+    navigator.mimeTypes["application/x-shockwave-flash"] = true;
+
+    // Test that we don't have flash
+    test.equal(zeroClipboard.detectFlashSupport(), true);
 
     test.done();
   },