+++ docs/releases.md
@@ -1,3 +1,8 @@
+### ZeroClipboard 1.1.0
+
+* [FEATURE] Replace dom selector with simple sizzler query selector
+* [FEATURE] Indicate via an event when there is no flash or flash is wrong version
+
 ### ZeroClipboard 1.0.9
 
 * [FEATURE] Use `data-clipboard-text` as default text #35